refactor: interface extracted from AppUserService, implementation created as AppUserServiceImpl
